Moves parallel_sort to its own example folder.
