+++ source/examples/parallel_sort/CMakeLists.txt
@@ -3,15 +3,18 @@
 # External dependencies
 # 
 
-find_package(${META_PROJECT_NAME} REQUIRED HINTS "${CMAKE_CURRENT_SOURCE_DIR}/../../../")
+# find_package(THIRDPARTY REQUIRED)
+
 
 # 
 # Executable name and options
 # 
 
 # Target name
-set(target yats-test)
-message(STATUS "Test ${target}")
+set(target parallel_sort)
+
+# Exit here if required dependencies are not met
+message(STATUS "Example ${target}")
 
 
 # 
@@ -19,13 +22,8 @@ message(STATUS "Test ${target}")
 # 
 
 set(sources
-    connection_test.cpp
-    identifier_test.cpp
     main.cpp
-    scheduler_test.cpp
-    taskconfigurator_test.cpp
-    util_test.cpp
-    yats_test.cpp
+    parallel_sort.h
 )
 
 
@@ -35,8 +33,12 @@ set(sources
 
 # Build executable
 add_executable(${target}
+    MACOSX_BUNDLE
     ${sources}
 )
+
+# Create namespaced alias
+add_executable(${META_PROJECT_NAME}::${target} ALIAS ${target})
 
 
 # 
@@ -58,6 +60,8 @@ target_include_directories(${target}
     PRIVATE
     ${DEFAULT_INCLUDE_DIRECTORIES}
     ${PROJECT_BINARY_DIR}/source/include
+	${CMAKE_CURRENT_SOURCE_DIR}
+	
 )
 
 
@@ -69,8 +73,8 @@ target_link_libraries(${target}
     PRIVATE
     ${DEFAULT_LIBRARIES}
     yats-include
-    gmock-dev
 )
+
 
 # 
 # Compile definitions
@@ -99,4 +103,25 @@ target_compile_options(${target}
 target_link_libraries(${target}
     PRIVATE
     ${DEFAULT_LINKER_OPTIONS}
+)
+
+
+#
+# Target Health
+#
+
+perform_health_checks(
+    ${target}
+    ${sources}
+)
+
+
+# 
+# Deployment
+# 
+
+# Executable
+install(TARGETS ${target}
+    RUNTIME DESTINATION ${INSTALL_BIN} COMPONENT examples
+    BUNDLE  DESTINATION ${INSTALL_BIN} COMPONENT examples
 )
